Trigger Fuction
This trigger assumes that the groups_id column in the match table references the groups_id column in the groups table. It updates the groups_position for the group associated with the newly inserted match record based on the number of goals scored by the home and away teams in the match. If the home team scored more goals, the groups_position is set to 1. Otherwise, it is set to 2.